requirements.txt
streamlit==1.35.0
moviepy==1.0.3
pytube==15.0.0
python-dotenv==1.0.1
groq==0.23.1
langchain==0.1.17
openai==1.24.0
httpx==0.23.3
tiktoken==0.6.0
pinecone-client==3.2.2
